requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0

# HTTP & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
urllib3>=2.0.0

# Security
cryptography>=41.0.0

# DNS & Domain
dnspython>=2.4.0
tldextract>=5.0.0
python-whois>=0.8.0

# ML & Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Caching
redis>=4.5.0
hiredis>=2.2.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Async
aiohttp>=3.8.0
asyncio>=3.4.3
